Add files via upload
Prediction of weather the patient gets readmitted to hospital or not based on his diabetic condition. We had the of hospital of US where the records of patient where there based on that by building the ML-Algorithms we have to predict will he gets admitted in the next 30 days or not. Can be deployed by building web application regarding this to predict weather patient is getting readmitted or not.